fixed previous commit

+++ src/math/matrix33.h
@@ -964,7 +964,7 @@ public:
     static Matrix33 rotationAroundZ(real angle);
     
     /// a rotation around one the axis X if `x==0`, Y if `x==1` or Z if `x==2`
-    static Matrix33 rotationAroundPrincipalAxis(unsigned x, real angle);
+    static Matrix33 rotationAroundPrincipalAxis(size_t x, real angle);
 
     /// return a rotation that transforms (1,0,0) into `vec` ( norm(vec) should be > 0 )
     static Matrix33 rotationToVector(const Vector3&);
